Extract all tools to /pages/tools
