+++ src/features/tools/employees/components/EmployeeFilter.tsx
@@ -1,10 +1,9 @@
 import styled from '@emotion/styled'
 import { SearchableDropdown } from '@hedvig-ui'
-import { Row } from 'features/tools/employees'
-import { CreateEmployee } from 'features/tools/employees/components/CreateEmployee'
 import React from 'react'
 import { Input } from 'semantic-ui-react'
 import { useAvailableEmployeeRolesQuery } from 'types/generated/graphql'
+import { CreateEmployee, Row } from './CreateEmployee'
 
 const StyledDropdown = styled(SearchableDropdown)`
   width: 12em;
