+++ src/pages/routes.tsx
@@ -1,12 +1,10 @@
 import { StandaloneMessage } from '@hedvig-ui'
-import { CampaignCodesPage } from 'features/tools/campaign-codes'
-import { ChargesPage } from 'features/tools/charges'
-import { EmployeesPage } from 'features/tools/employees'
-import { ItemizerPage } from 'features/tools/itemizer'
-import { NorwegianTariffCreatorPage } from 'features/tools/norwegian-tariff-editor'
-import { PerilsEditorPage } from 'features/tools/perils-editor'
-import { UnsignMemberPage } from 'features/tools/staging-tools/unsign-member-tool'
-import { SwitcherAutomationPage } from 'features/tools/switcher-automation'
+import { CampaignCodesPage } from 'pages/tools/campaign-codes'
+import { ItemizerPage } from 'pages/tools/itemizer'
+import { NorwegianTariffCreatorPage } from 'pages/tools/norwegian-tariff-creator'
+import { PerilsEditorPage } from 'pages/tools/perils-editor'
+import { SwitcherAutomationPage } from 'pages/tools/switcher-automation'
+import { UnsignMemberPage } from 'pages/tools/unsign-member'
 import React from 'react'
 import { Route, Switch } from 'react-router'
 import { ClaimDetails } from './claims/[claimId]/members/[memberId]'
@@ -16,6 +14,8 @@ import { MembersPage } from './members'
 import { MemberPage } from './members/[memberId]/[tab]'
 import { QuestionsPage } from './questions'
 import { ToolsPage } from './tools'
+import { ChargesPage } from './tools/charges'
+import { EmployeesPage } from './tools/employees'
 
 export const Routes: React.FC = () => {
   return (
