+++ src/pages/tools/itemizer.tsx
@@ -1,0 +1,50 @@
+import { MainHeadline } from '@hedvig-ui'
+import { TableInput } from 'features/tools/itemizer/TableInput'
+import React from 'react'
+import {
+  useInsertItemCategoriesMutation,
+  useInsertValuationRulesMutation,
+} from 'types/generated/graphql'
+
+export const ItemizerPage: React.FC = () => {
+  const [insertItemCategories] = useInsertItemCategoriesMutation()
+  const [insertValuationRules] = useInsertValuationRulesMutation()
+
+  return (
+    <>
+      <MainHeadline>Itemizer</MainHeadline>
+      <TableInput
+        title="Valuation rules"
+        headers={[
+          'Name',
+          'Age',
+          'Type of Contract',
+          'Item Family',
+          'Item Type',
+          'Valuation Type',
+          'Depreciation',
+        ]}
+        onSubmit={(data, setValidity, setResetRequired) =>
+          insertValuationRules({
+            variables: { request: { valuationRulesString: data } },
+          }).then(({ data: response }) => {
+            setResetRequired(true)
+            setValidity(response?.insertValuationRules ?? [])
+          })
+        }
+      />
+      <TableInput
+        title="Item categories"
+        headers={['Family', 'Type', 'Company', 'Brand', 'Model']}
+        onSubmit={(data, setValidity, setResetRequired) =>
+          insertItemCategories({
+            variables: { request: { itemCategoriesString: data } },
+          }).then(({ data: response }) => {
+            setResetRequired(true)
+            setValidity(response?.insertItemCategories ?? [])
+          })
+        }
+      />
+    </>
+  )
+}
